+++ demo/app/pages/+default/home/home.component.ts
@@ -2,7 +2,7 @@ import {Component, ChangeDetectionStrategy} from '@angular/core';
 import {ComponentRedirectRoute, ComponentRoute} from '@anglr/common/router';
 import {DebugDataCopyClickModule} from '@anglr/common/material';
 import {WithScrollableCssClass} from '@anglr/common';
-import {SelectComponent} from '@anglr/select';
+import {SelectModule} from '@anglr/select';
 
 /**
  * Home component
@@ -14,7 +14,7 @@ import {SelectComponent} from '@anglr/select';
     standalone: true,
     imports:
     [
-        SelectComponent,
+        SelectModule,
         DebugDataCopyClickModule,
     ],
     changeDetection: ChangeDetectionStrategy.OnPush