+++ src/misc/classes/index.ts
@@ -1,1 +1,2 @@
 export * from './selectBus/selectBus';
+export * from './selectPluginInstances/selectPluginInstances';